adding info split-panel for cells
organizing code into module
updating some css
