+++ object_database/web/devtools/cells_panel.html
@@ -8,6 +8,7 @@
   </head>
   <body>
       <div id="main"></div>
+      <div id="cell-info"> some cells data here</div>
   </body>
-  <script src="cell_panel.js"></script>
+  <script type="module" src="js/cell_panel.js"></script>
 </html>
